requirements.txt
django
djangorestframework
gunicorn
django-heroku
whitenoise
dj-database-url
black